delete_branch_after_merge: false

+++ src/pull/merged.js
@@ -1,13 +1,14 @@
-// const getConfig = require('../config')
+const getConfig = require('../config')
 
 module.exports = () => async context => {
   const thread = context.payload.pull_request
 
   if (thread.merged === true) {
-    // const config = await getConfig(context)
-    // TODO check config if branch should be deleted
+    const config = await getConfig(context)
 
-    const ref = `heads/${thread.head.ref}`
-    return context.github.gitdata.deleteRef(context.repo({ ref }))
+    if (config.delete_branch_after_merge) {
+      const ref = `heads/${thread.head.ref}`
+      return context.github.gitdata.deleteRef(context.repo({ ref }))
+    }
   }
 }
